+++ classes/Ninja/DatabaseTable.php
@@ -1,10 +1,12 @@
 <?php
+namespace Ninja;
+
 class DatabaseTable {
 	private $pdo;
 	private $table;
 	private $primaryKey;
 
-	public function __construct(PDO $pdo, string $table, string $primaryKey) {
+	public function __construct(\PDO $pdo, string $table, string $primaryKey) {
 		$this->pdo = $pdo;
 		$this->table = $table;
 		$this->primaryKey = $primaryKey;
