+++ classes/Ninja/EntryPoint.php
@@ -1,4 +1,6 @@
 <?php
+namespace Ninja;
+
 class EntryPoint {
 	private $route;
 	private $routes;
@@ -20,7 +22,7 @@ class EntryPoint {
 		extract($variables);
 
 		ob_start();
-		include  __DIR__ . '/../templates/' . $templateFileName;
+		include  __DIR__ . '/../../templates/' . $templateFileName;
 
 		return ob_get_clean();
 	}
@@ -38,6 +40,6 @@ class EntryPoint {
 			$output = $this->loadTemplate($page['template']);
 		}
 
-		include  __DIR__ . '/../templates/layout.html.php';
+		include  __DIR__ . '/../../templates/layout.html.php';
 	}
 }
