+++ includes/autoload.php
@@ -1,6 +1,9 @@
 <?php
 function autoloader($className) {
-	$file = __DIR__ . '/../classes/' . $className . '.php';
+	$fileName = str_replace('\\', '/', $className) . '.php';
+
+	$file = __DIR__ . '/../classes/' . $fileName;
+	
 	include $file;
 }
