+++ public/index.php
@@ -4,7 +4,7 @@ try {
 	
 	$route = ltrim(strtok($_SERVER['REQUEST_URI'], '?'), '/');
 
-	$entryPoint = new EntryPoint($route, new IjdbRoutes());
+	$entryPoint = new \Ninja\EntryPoint($route, new \Ijdb\IjdbRoutes());
 	$entryPoint->run();
 }
 catch (PDOException $e) {
